oud/requirements copy.txt
yfinance
pandas==1.5.3
pandas_ta
pandas_datareader
statsmodels
matplotlib
scikit-learn
PyPortfolioOpt
jupyterlab
arch
Flask 
flask_cors
alpaca-py
alpaca
requests
python-dotenv
streamlit
numpy
matplotlib
